+++ dimidium/middleend/archGen/ArchNode.py
@@ -79,9 +79,9 @@ class ArchNode(object):
                'selected_hw_type': repr(self.selected_hw_type),
                'estimations': {'comp_util%': self.used_comp_util_share*100, 'mem_util%': self.used_mem_util_share*100,
                                # 'used_perf%': self.used_perf_share*100},
-                               'req_iter_hz': self.req_iter_hz, 'used_iter_hz': self.used_iter_hz,
-                               'impl_Gflop_eqiv': self.used_perf_F / gigaU,
-                               'max_Gflop_based_on_impl_eqiv': self.max_perf_iter_based / gigaU},
+                               'req_iter_hz': self.req_iter_hz, 'used_iter_hz': float(self.used_iter_hz),
+                               'impl_Gflop_eqiv': float(self.used_perf_F / gigaU),
+                               'max_Gflop_based_on_impl_eqiv': float(self.max_perf_iter_based / gigaU)},
                'over_utilized_node': self.over_utilized_node,
                'blocks': [], 'engineContainers': [],
                'bricks': {}}
